requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<3.16
Pillow>=10.0.0,<11.0.0
django-jet==1.0.8
qrcode>=7.4.2,<8.0.0
